+++ tests/unit/controllers/v0.3/IlsClient.test.js
@@ -564,6 +564,7 @@ describe("IlsClient", () => {
       policy,
       ilsClient: IlsClient({}),
       varFields: [{ fieldTag: "x", content: "DEPENDENT OF 1234" }],
+      acceptTerms: true,
     });
 
     it("returns an ILS-ready patron object", async () => {
@@ -651,6 +652,7 @@ describe("IlsClient", () => {
       address,
       policy,
       ilsClient: IlsClient({}),
+      acceptTerms: true,
     });
     // Mock that the ptype was added to the card.
     card.setPtype();